refactor(requests): move enums to enums folder
