+++ backend/src/requests/dto/update-request.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty, PartialType } from '@nestjs/swagger';
 import { IsEnum } from 'class-validator';
-import { RequestStatus } from '../requests.enum';
+import { RequestStatus } from '../enums/request.enums';
 import { CreateRequestDto } from './create-request.dto';
 
 export class UpdateRequestDto extends PartialType(CreateRequestDto) {
