+++ backend/src/requests/enums/request.enums.ts
@@ -1,1 +1,9 @@
-export { RequestStatus } from '../requests.enum';
+export enum RequestStatus {
+  PENDING = 'pending',
+  ACCEPTED = 'accepted',
+  REJECTED = 'rejected',
+  IN_PROGRESS = 'inProgress',
+  DONE = 'done',
+  APPROVED = 'APPROVED',
+  CANCELLED = 'CANCELLED',
+}
