--- backend/src/requests/requests.enum.ts
@@ -1,9 +1,0 @@
-export enum RequestStatus {
-  PENDING = 'pending',
-  ACCEPTED = 'accepted',
-  REJECTED = 'rejected',
-  IN_PROGRESS = 'inProgress',
-  DONE = 'done',
-  APPROVED = 'APPROVED',
-  CANCELLED = 'CANCELLED',
-}
